+++ src/store/api-actions.ts
@@ -1,32 +1,27 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
 import { AxiosInstance } from 'axios';
-import { AppDispatch, RootState } from '.';
-import { AuthorizationStatus } from '../const';
+import { AppDispatch, RootState, store } from '.';
+import { APIRoute, AuthorizationStatus, TIMEOUT_SHOW_ERROR } from '../const';
 import { saveToken, dropToken } from '../services/token';
-import { requireAuthorization, setOffers } from '../features/sorting-offers-by-cities';
-import { Offer } from '../types/types';
+import { requireAuthorization, setError, setOffers } from '../features/sorting-offers-by-cities';
+import { AuthData, Offer, UserData } from '../types/types';
 
-export type AuthData = {
-  login: string;
-  password: string;
-};
-export type UserData = {
-  id: number;
-  email: string;
-  token: string;
-};
-export enum APIRoute {
-  Offers = '/offers',
-  Login = '/login',
-  Logout = '/logout',
-}
+export const clearErrorAction = createAsyncThunk(
+  '/clearError',
+  () => {
+    setTimeout(
+      () => store.dispatch(setError(null)),
+      TIMEOUT_SHOW_ERROR,
+    );
+  },
+);
 
-export const fetchQuestionAction = createAsyncThunk<void, undefined, {
+export const fetchOffersAction = createAsyncThunk<void, undefined, {
   dispatch: AppDispatch;
   state: RootState;
   extra: AxiosInstance;
 }>(
-  'data/fetchQuestions',
+  '/offers',
   async (_arg, {dispatch, extra: api}) => {
     const {data} = await api.get<Offer[]>(APIRoute.Offers);
     dispatch(setOffers(data));
@@ -38,7 +33,7 @@ export const checkAuthAction = createAsyncThunk<void, undefined, {
   state: RootState;
   extra: AxiosInstance;
 }>(
-  'user/checkAuth',
+  '/checkAuth',
   async (_arg, {dispatch, extra: api}) => {
     try {
       await api.get(APIRoute.Login);
@@ -54,7 +49,7 @@ export const loginAction = createAsyncThunk<void, AuthData, {
   state: RootState;
   extra: AxiosInstance;
 }>(
-  'user/login',
+  '/login',
   async ({login: email, password}, {dispatch, extra: api}) => {
     const {data: {token}} = await api.post<UserData>(APIRoute.Login, {email, password});
     saveToken(token);
@@ -67,7 +62,7 @@ export const logoutAction = createAsyncThunk<void, undefined, {
   state: RootState;
   extra: AxiosInstance;
 }>(
-  'user/logout',
+  '/logout',
   async (_arg, {dispatch, extra: api}) => {
     await api.delete(APIRoute.Logout);
     dropToken();
